Сделал дз 9-1
