+++ index.html
@@ -120,7 +120,7 @@
                     <input class="text__hashtags" name="hashtags" placeholder="#ХэшТег">
                   </div>
                   <div class="img-upload__field-wrapper">
-                    <textarea class="text__description" name="description" placeholder="Ваш комментарий..."></textarea>
+                    <textarea class="text__description" name="description" data-pristine-maxlength="140" data-pristine-maxlength-message="Комментарий не может быть больше 140 символов." placeholder="Ваш комментарий..."></textarea>
                   </div>
                 </fieldset>
 
@@ -250,5 +250,6 @@
 <script src="js/main.js" type="module"></script>
 <script src="js/functions.js"></script>
 <script src="vendor/pristine/pristine.min.js"></script>
+<script src="vendor/nouislider/nouislider.js"></script>
 </body>
 </html>
