+++ js/main.js
@@ -1,6 +1,5 @@
 import { generatePhotos } from './data.js';
 import { renderGallery } from './gallery.js';
-import './upload-overlay.js';
-import './form.js';
+import './upload-form.js';
 
 renderGallery(generatePhotos());
